@@ -1,5 +1,5 @@
 import { Component, OnDestroy, OnInit } from '@angular/core';
-import { FormControl } from '@angular/forms';
+import { UntypedFormControl } from '@angular/forms';
 import { ReplaySubject, Subject } from 'rxjs';
 import {debounceTime, delay, tap, filter, map, takeUntil} from 'rxjs/operators';
 
@@ -17,10 +17,10 @@ export class ServerSideSearchExampleComponent implements OnInit, OnDestroy {
   protected banks: Bank[] = BANKS;
 
   /** control for the selected bank for server side filtering */
-  public bankServerSideCtrl: FormControl = new FormControl();
+  public bankServerSideCtrl: UntypedFormControl = new UntypedFormControl();
 
   /** control for filter for server side. */
-  public bankServerSideFilteringCtrl: FormControl = new FormControl();
+  public bankServerSideFilteringCtrl: UntypedFormControl = new UntypedFormControl();
 
   /** indicate search operation is in progress */
   public searching = false;
